# docs/swagger.yaml
basePath: /v1
definitions:
  main.ChangePasswordPayload:
    properties:
      password:
        maxLength: 72
        minLength: 3
        type: string
    required:
    - password
    type: object
...
  main.CreateUpdateUsernamePayload:
    properties:
      email:
        maxLength: 255
        type: string
      username:
        maxLength: 255
        type: string
    type: object
...
  store.User:
    properties:
      coin:
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      token_version:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  swagger.EnvelopeError:
    properties:
      error:
        type: string
    type: object
  swagger.EnvelopeString:
    properties:
      data:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: RESTful API for an online novel platform — read, publish, and manage
    digital stories.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Govel API
paths:
...
  /users/:
    get:
      consumes:
      - application/json
      description: Retrieve the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User profile data
          schema:
            $ref: '#/definitions/store.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update user profile, including username and/or email
      parameters:
      - description: Updated user profile data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.CreateUpdateUsernamePayload'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user profile data
          schema:
            $ref: '#/definitions/store.User'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - users
  /users/{userID}:
    get:
      consumes:
      - application/json
      description: Retrieve the user profile by user ID
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User profile data
          schema:
            $ref: '#/definitions/store.User'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
  /users/activate/{token}:
    put:
      consumes:
      - application/json
      description: Activate user account using a valid token
      parameters:
      - description: Activation token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: User account activated
          schema:
            type: ""
        "404":
          description: invalid token
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      summary: Activate user account
      tags:
      - users
  /users/change-password:
    patch:
      consumes:
      - application/json
      description: Change the password of the currently authenticated user
      parameters:
      - description: New password
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.ChangePasswordPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Password change success message
          schema:
            $ref: '#/definitions/swagger.EnvelopeString'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - users
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"